I've been haunted by this problem for 2 weeks.
Messages received over HTTP do *NOT* have a guild_id set. Returning , the previous behaviour of this function, could almost never fail and if it did it was because you had a DM channel.

HOWEVER! Messages received over the Gateway **DO** have a guild_id set. I'm going to prefer the legacy behavior but keep the new behavior in as a fallback because I do not understand why this change was made but will trust it was made in good faith.